+++ assets/js/main.js
@@ -7,14 +7,3 @@
 
 })(jQuery);
 
-/*
- * Add wrap and class embed-responsive videos
- */
-(function($) {
-
-	var $iframes = $('.entry-content p:has(iframe)');
-	var $iframesContent = $('.entry-content p > iframe');
-	$iframes.wrap('<div class="embed-responsive embed-responsive-16by9"></div>');
-	$iframesContent.addClass('embed-responsive-item');
-
-})(jQuery);