+++ config/responsive-menus.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Moare Genesis.
+ *
+ * @package moaregenesis
+ * @author  Antonio
+ * @since   1.0.7
+ * @license GPL-2.0+
+ * @link    http://studiomoare.com/
+ * @version 1.0.7
+ */
+
+/**
+ * Genesis responsive menus settings. (Requires Genesis 3.0+.)
+ */
+return [
+	'script' => [
+		'menuClasses' => [
+			'others' => [ '.nav-primary' ],
+		],
+	],
+	'extras' => [
+		'media_query_width' => '992px',
+	],
+];